+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master (unreleased)
 
+- Support PostgreSQL schemas for `#fast_count`
+
 ## 0.1.0 (2023-04-26)
 
 - First release